refactor(firmware-update): move OTA param building to notifier layer
Address code review feedback:
- Fix: Remove PII (MAC/IP) from log by only logging URI path
- Fix: Move OTA check param building logic from View to Notifier
- buildOtaCheckParams(), _formatMacAddress(), _parseHardwareVersion()
- View now only calls notifier methods, no business logic
